+++ src/main/java/com/flab/foodeats/application/menu/EssentialMenuTarget.java
@@ -4,8 +4,8 @@ import com.flab.foodeats.domain.menu.Menu;
 
 public class EssentialMenuTarget {
 
-	private int shopId;
-	private int menuId;
+	private long shopId;
+	private long menuId;
 	private String menuName;
 	private int menuPrice;
 	private String menuConfiguration;
@@ -15,7 +15,7 @@ public class EssentialMenuTarget {
 	public EssentialMenuTarget() {
 	}
 
-	public EssentialMenuTarget(int shopId, String menuName, int menuPrice, String menuConfiguration,
+	public EssentialMenuTarget(long shopId, String menuName, int menuPrice, String menuConfiguration,
 		String menuExplanation, String menuGroup) {
 		this.shopId = shopId;
 		this.menuName = menuName;
@@ -25,7 +25,7 @@ public class EssentialMenuTarget {
 		this.menuGroup = menuGroup;
 	}
 
-	public EssentialMenuTarget(int shopId, int menuId, String menuName, int menuPrice, String menuConfiguration,
+	public EssentialMenuTarget(long shopId, long menuId, String menuName, int menuPrice, String menuConfiguration,
 		String menuExplanation, String menuGroup) {
 		this.shopId = shopId;
 		this.menuId = menuId;
@@ -36,11 +36,11 @@ public class EssentialMenuTarget {
 		this.menuGroup = menuGroup;
 	}
 
-	public int getShopId() {
+	public long getShopId() {
 		return shopId;
 	}
 
-	public int getMenuId() {
+	public long getMenuId() {
 		return menuId;
 	}
