+++ src/main/java/com/flab/foodeats/application/menu/port/MenuOptionService.java
@@ -3,12 +3,13 @@ package com.flab.foodeats.application.menu.port;
 import java.util.List;
 
 import com.flab.foodeats.api.menu.OptionRequest;
+import com.flab.foodeats.application.menu.OptionTarget;
 
 public interface MenuOptionService {
 
-	void registerMenuOption(Long shopId, int menuId, List<OptionRequest> optionRequests, String userId);
+	void registerMenuOption(long shopId, long menuId, List<OptionRequest> optionRequests, String userId);
 
-	void updateMenuOption(Long shopId, OptionTarget optionTarget, String userId);
+	void updateMenuOption(long shopId, OptionTarget optionTarget, String userId);
 
-	void deleteMenuOption(Long shopId, int menuOptionId, String userId);
+	void deleteMenuOption(long shopId, long menuOptionId, String userId);
 }
