+++ src/main/java/com/flab/foodeats/application/menu/port/MenuService.java
@@ -8,5 +8,5 @@ public interface MenuService {
 
 	void modifyMenu(EssentialMenuTarget essentialMenuTarget, String userId);
 
-	void deleteMenu(int shopId, int menuId, String UserId);
+	void deleteMenu(long shopId, long menuId, String UserId);
 }
